LiDARGen/requirements.txt
absl-py>=0.10.0
addict>=2.4.0
anyio>=3.3.4
argon2-cffi>=21.1.0
backcall>=0.2.0
bleach>=4.1.0
clang>=5.0
contextlib2>=21.6.0
debugpy>=1.5.1
defusedxml>=0.7.1
deprecation>=2.1.0
dill>=0.3.4
dm-tree>=0.1.6
entrypoints>=0.3
flatbuffers>=1.12
grpcio>=1.41.0
h5py>=3.1.0
ipykernel>=6.4.2
ipython>=7.28.0
ipython-genutils>=0.2.0
ipywidgets>=7.6.5
jedi>=0.18.0
joblib>=1.1.0
json5>=0.9.6
jsonschema>=4.1.2
jupyter-client>=7.0.6
jupyter-core>=4.8.1
jupyter-packaging>=0.11.0
jupyter-server>=1.11.1
jupyterlab>=3.2.1
jupyterlab-pygments>=0.1.2
jupyterlab-server>=2.8.2
jupyterlab-widgets>=1.0.2
keras>=2.6.0
markupsafe>=2.0.1
matplotlib-inline>=0.1.3
mistune>=0.8.4
ml-collections>=0.1.0
nbclassic>=0.3.3
nbclient>=0.5.4
nbconvert>=6.2.0
nbformat>=5.1.3
nest-asyncio>=1.5.1
notebook>=6.4.5
numpy>=1.19.5
open3d>=0.13.0
opencv-python>=4.5.4.58
packaging>=21.0
pandocfilters>=1.5.0
parso>=0.8.2
pexpect>=4.8.0
pickleshare>=0.7.5
prometheus-client>=0.11.0
promise>=2.3
prompt-toolkit>=3.0.21
ptyprocess>=0.7.0
pygments>=2.10.0
pyrsistent>=0.18.0
pyyaml>=6.0
pyzmq>=22.3.0
requests-unixsocket>=0.2.0
scikit-learn>=1.0.1
send2trash>=1.8.0
six>=1.15.0
sniffio>=1.2.0
tensorboard>=2.7.0
tensorboard-data-server>=0.6.1
tensorflow>=2.6.0
tensorflow-addons>=0.12.0
tensorflow-datasets>=3.2.0
tensorflow-gan>=2.0.0
tensorflow-hub>=0.12.0
tensorflow-io>=0.21.0
tensorflow-io-gcs-filesystem>=0.21.0
tensorflow-metadata>=1.2.0
tensorflow-probability>=0.14.1
terminado>=0.12.1
testpath>=0.5.0
threadpoolctl>=3.0.0
tomlkit>=0.7.2
torch-tb-profiler>=0.3.1
traitlets>=5.1.1
typeguard>=2.13.0
wcwidth>=0.2.5
webencodings>=0.5.1
websocket-client>=1.2.1
widgetsnbextension>=3.5.1